feat: create new transaction response that has the earliest date along with the transactions

+++ wallet/src/main/kotlin/net/thechance/wallet/api/dto/transaction/TransactionProjectionMapper.kt
@@ -13,6 +13,6 @@ fun TransactionProjection.toResponse(): TransactionResponse {
         status = status,
         type = TransactionType.valueOf(type),
         createdAt = createdAt,
-        amount = amount,
+        amount = amount.toDouble(),
     )
 }
